feat: stop parameter and MessageType
Parse the raw u8 into a MessageType for easier recognition and pattern matching.

Allow providing a parameter when running the program that specifies whether the program should stop when it encounters the last message or whether the program should continue to scan the memory for new messages and display them as they are received.