refactor: do not implement IntoIterator
IntoIterator has the side effect of dropping the original object, which is expected, but that also calls stop(). That effectively makes the IntoIterator pattern useless, as the into_iter() call will immediately drop, thereby stopping, the reader.

I cannot in good conscience remove the Drop implementation, since that would not be in the spirit of good Rust practices.